Move regepx range extractor to utils namespace
